git-svn-id: http://plugins.svn.wordpress.org/wp-e-commerce@70822 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ trunk/wp-shopping-cart/display-items.php
@@ -7,8 +7,11 @@ $current_user = wp_get_current_user();
 
 $closed_postboxes = (array)get_usermeta( $current_user->ID, 'closedpostboxes_products');
 
-  $variations_processor = new nzshpcrt_variations;
+$variations_processor = new nzshpcrt_variations;
 
+$flash = true;
+if ( false !== strpos(strtolower($_SERVER['HTTP_USER_AGENT']), 'mac') && apache_mod_loaded('mod_security') )
+	$flash = false;
 function topcategorylist() {
 	global $wpdb,$category_data;
 	$siteurl = get_option('siteurl');
@@ -264,17 +267,15 @@ if($_POST['submit_action'] == 'add') {
 			
 			$variations_processor = new nzshpcrt_variations;
 			if($_POST['variations'] != null) {
-			
-        foreach((array)$_POST['variations'] as $variation_id => $state) {
-          $variation_id = (int)$variation_id;
-          if($state == 1) {
-            $variation_values = $variations_processor->falsepost_variation_values($variation_id);
-            $variations_processor->add_to_existing_product($product_id,$variation_values);
-          }
-        }
+				foreach((array)$_POST['variations'] as $variation_id => $state) {
+					$variation_id = (int)$variation_id;
+					if($state == 1) {
+						$variation_values = $variations_processor->falsepost_variation_values($variation_id);
+						$variations_processor->add_to_existing_product($product_id,$variation_values);
+					}
+				}
 			}
 
-				
 			if($_POST['variation_priceandstock'] != null) {
 				$variations_processor->update_variation_values($product_id, $_POST['variation_priceandstock']);
 	// 			  exit("<pre>".print_r($_POST,true)."</pre>");
@@ -1386,24 +1387,30 @@ echo "        </div>";
 	<table width='100%'>
     <tr>
       <td>
-			<button id="add-product-image" name="add-image" class="button-secondary" type="button"><small>Add New Image</small></button>
-        <?php// echo TXT_WPSC_PRODUCTIMAGE;?><!--:-->
-	</td>
-	<!-- <td>
-        <input type='file' name='image' value='' />
-	</td>-->
-    </tr>
-    
+      	<?php
+      	if ($flash)
+		echo '<button id="add-product-image" name="add-image" class="button-secondary" type="button"><small>Add New Image</small></button>';
+        else {
+		echo TXT_WPSC_PRODUCTIMAGE.":";
+		echo "</td>
+		<td>
+			<input type='file' name='image' value='' />
+		</td>
+	</tr>";
+    }
+    ?>
     <tr>
       <!--<td></td>--><td>
       <table>
 	<ul id='gallery_list'></ul>
   <?php
   // pe.{ & table opening above
-  if(function_exists("getimagesize") && is_numeric(get_option('product_image_height')) && is_numeric(get_option('product_image_width')))
-    {
+	
+	if ($flash) {
+	 } else {
+  if(function_exists("getimagesize") && is_numeric(get_option('product_image_height')) && is_numeric(get_option('product_image_width'))) {
     ?>
-<!--      <tr>
+      <tr>
         <td>
       <input type='radio' name='image_resize' value='0' id='add_image_resize0' class='image_resize' onclick='hideOptionElement(null, "image_resize0");' /> <label for='add_image_resize0'><?php echo TXT_WPSC_DONOTRESIZEIMAGE; ?></label>
         </td>
@@ -1412,42 +1419,42 @@ echo "        </div>";
         <td>
           <input type='radio' checked='true' name='image_resize' value='1' id='add_image_resize1' class='image_resize' onclick='hideOptionElement(null, "image_resize1");' /> <label for='add_image_resize1'><?php echo TXT_WPSC_USEDEFAULTSIZE;?> <?php echo "(<abbr title='".TXT_WPSC_SETONSETTINGS."'>".get_option('product_image_height') ."&times;".get_option('product_image_width')."px</abbr>)"; ?></label>
         </td>
-      </tr>-->
+      </tr>
     <?php  
     $default_size_set = true;
     }
   
-//   if(function_exists("getimagesize"))
-//     {
-//     ?>
-<!--      <tr>
-//         <td>
-//           <input type='radio' name='image_resize' value='2'id='add_image_resize2' class='image_resize'  onclick='hideOptionElement("heightWidth", "image_resize2");' />
-//       <label for='add_image_resize2'><?php echo TXT_WPSC_USESPECIFICSIZE; ?> </label>        
-//           <div id='heightWidth' style='display: none;'>
-//         <input type='text' size='4' name='width' value='' /><label for='add_image_resize2'><?php echo TXT_WPSC_PXWIDTH;?></label>
-//         <input type='text' size='4' name='height' value='' /><label for='add_image_resize2'><?php echo TXT_WPSC_PXHEIGHT; ?> </label>
-//       </div>
-//         </td>
-//       </tr>
-//       <tr>
-//       <td>
-//         <input type='radio' name='image_resize' value='3' id='add_image_resize3' class='image_resize' onclick='hideOptionElement("browseThumb", "image_resize3");' />
-//         <label for='add_image_resize3'><?php echo TXT_WPSC_SEPARATETHUMBNAIL; ?></label><br />
-//         <div id='browseThumb' style='display: none;'>
-//           <input type='file' name='thumbnailImage' value='' />
-//         </div>
-//       </td>
-//     </tr>
-//     -->
+  if(function_exists("getimagesize"))
+    {
+    ?>
+     <tr>
+        <td>
+           <input type='radio' name='image_resize' value='2'id='add_image_resize2' class='image_resize'  onclick='hideOptionElement("heightWidth", "image_resize2");' />
+       <label for='add_image_resize2'><?php echo TXT_WPSC_USESPECIFICSIZE; ?> </label>        
+           <div id='heightWidth' style='display: none;'>
+         <input type='text' size='4' name='width' value='' /><label for='add_image_resize2'><?php echo TXT_WPSC_PXWIDTH;?></label>
+        <input type='text' size='4' name='height' value='' /><label for='add_image_resize2'><?php echo TXT_WPSC_PXHEIGHT; ?> </label>
+       </div>
+         </td>
+       </tr>
+       <tr>
+       <td>
+         <input type='radio' name='image_resize' value='3' id='add_image_resize3' class='image_resize' onclick='hideOptionElement("browseThumb", "image_resize3");' />
+         <label for='add_image_resize3'><?php echo TXT_WPSC_SEPARATETHUMBNAIL; ?></label><br />
+         <div id='browseThumb' style='display: none;'>
+           <input type='file' name='thumbnailImage' value='' />
+         </div>
+       </td>
+     </tr>
     <?php
-//     }
+        }
     
-//     if(function_exists('add_multiple_image_form')) {
-//       echo add_multiple_image_form("add_");
-//       }
+    if(function_exists('add_multiple_image_form')) {
+      echo add_multiple_image_form("add_");
+      }
 	if(function_exists('gold_shpcrt_install')) {
 		echo "<input type='hidden' value='1' id='gold_present'>";
+	}
 	}
   ?>
         </table>
